feat: update VerificationCodeModel to use uint64 for ID and UserID, and add DeleteByUserID method in VerificationCodeRepository

+++ internal/modules/identity/model/verification_code_model.go
@@ -5,8 +5,8 @@ import (
 )
 
 type VerificationCodeModel struct {
-	ID        uint `gorm:"primarykey"`
-	UserID    uint
+	ID        uint64 `gorm:"primarykey"`
+	UserID    uint64
 	Code      string
 	ExpiresAt time.Time
 	CreatedAt time.Time
